Move custom symbols to worker_ext.go
