+++ miner/worker.go
@@ -25,7 +25,6 @@ import (
 	"sync/atomic"
 	"time"
 
-	"github.com/ava-labs/avalanchego/utils/units"
 	"github.com/ava-labs/coreth/consensus"
 	"github.com/ava-labs/coreth/consensus/misc/eip4844"
 	"github.com/ava-labs/coreth/core"
@@ -76,10 +75,6 @@ const (
 
 	// staleThreshold is the maximum depth of the acceptable stale block.
 	staleThreshold = 7
-
-	// Leaves 256 KBs for other sections of the block (limit is 2MB).
-	// This should suffice for atomic txs, proposervm header, and serialization overhead.
-	targetTxsSize = 1792 * units.KiB
 )
 
 var (
@@ -733,12 +728,6 @@ func (w *worker) resultLoop() {
 	}
 }
 
-// withConcurrentWorkers is a simple wrapper to reduce diffs in the [worker.makeEnv] method
-// compared to the Geth codebase, especially to have the state argument named `state`.
-func withConcurrentWorkers(numPrefetchers int) state.PrefetcherOption {
-	return state.WithConcurrentWorkers(numPrefetchers)
-}
-
 // makeEnv creates a new environment for the sealing block.
 func (w *worker) makeEnv(parent *types.Header, header *types.Header, coinbase common.Address, predicateContext *precompileconfig.PredicateContext) (*environment, error) {
 	// Retrieve the parent state to execute on top and start a prefetcher for
@@ -1156,15 +1145,6 @@ func (w *worker) fillTransactions(interrupt *atomic.Int32, env *environment) err
 		}
 	}
 	return nil
-}
-
-// isDurango is a small helper function with its main purpose being to reduce diffs
-// in the [worker.generateWork] method compared to the Geth codebase. Notably, it allows
-// to keep the argument name as `params` and not conflict with the `params` package.
-func (w *worker) isDurango(work *environment) bool {
-	rules := w.chainConfig.Rules(work.header.Number, params.IsMergeTODO, work.header.Time)
-	rulesExtra := params.GetRulesExtra(rules)
-	return rulesExtra.IsDurango
 }
 
 // generateWork generates a sealing block based on the given parameters.
